Added Borda method

+++ routes/web.php
@@ -52,35 +52,36 @@ Route::group(['middleware' => ['auth']],function () {
         Route::post('/store', [UserCategoriesController::class, 'store'])->name('store-categories');
         Route::get('/edit', [UserCategoriesController::class, 'edit'])->name('edit-categories');
         Route::post('/update', [UserCategoriesController::class, 'update'])->name('update-categories');
-        Route::delete('/destroy/{id}', [UserCategoriesController::class, 'destroy'])->name('destroy-categories');
+        // Route::delete('/destroy/{id}', [UserCategoriesController::class, 'destroy'])->name('destroy-categories');
     });
 
         Route::prefix('category')->group(function() {
-            Route::get('/', [CategoryController::class, 'index'])->name('category');
-            Route::get('/create', [CategoryController::class, 'create'])->name('create-category');
-            Route::get('/store', [CategoryController::class, 'store'])->name('store-category');
-            Route::get('/edit', [CategoryController::class, 'edit'])->name('store-category');
-            Route::get('/update', [CategoryController::class, 'update'])->name('update-category');
+            Route::get('/', [CategoryController::class, 'index'])->middleware('permission:read_category')->name('category');
+            Route::get('/create', [CategoryController::class, 'create'])->middleware('permission:create_category')->name('create-category');
+            Route::post('/store', [CategoryController::class, 'store'])->middleware('permission:create_category')->name('store-category');
+            Route::get('/edit/{id}', [CategoryController::class, 'edit'])->middleware('permission:update_category')->name('edit-category');
+            Route::post('/update/{id}', [CategoryController::class, 'update'])->middleware('permission:update_category')->name('update-category');
+            Route::delete('/destroy/{id}', [CategoryController::class, 'update'])->middleware('permission:update_category')->name('update-category');
         });
 
         Route::prefix('school')->group(function () {
-            Route::get('/', [SchoolController::class, 'index'])->middleware('permission:read_school')->name('school');
+            Route::get('/', [SchoolController::class, 'index'])->name('school');
             ROute::get('/show/{id}', [SchoolDetailController::class, 'show'])->name("show-school");
-            Route::get('/create', [SchoolController::class, 'create'])->name('create-school');
-            Route::get('/create-detail/{id}', [SchoolDetailController::class, 'create'])->name('create-detail');
-            Route::post('/store', [SchoolController::class, 'store'])->name('store-school');
-            Route::post('/store-detail/{id}', [SchoolDetailController::class, 'store'])->name('store-detail');
-            Route::get('/edit/{id}', [SchoolController::class, 'edit'])->name('edit-school');
-            Route::get('/edit-detail/{id}', [SchoolDetailController::class, 'edit'])->name('edit-detail');
-            Route::post('/update/{id}', [SchoolController::class, 'update'])->name('update-school');
-            Route::post('/update-detail/{id}', [SchoolDetailController::class, 'update'])->name('update-detail');
-            Route::delete('/destroy/{id}', [SchoolController::class, 'destroy'])->name('destroy-school');
+            Route::get('/create', [SchoolController::class, 'create'])->middleware('permission:create_school')->name('create-school');
+            Route::get('/create-detail/{id}', [SchoolDetailController::class, 'create'])->middleware('permission:create_school')->name('create-detail');
+            Route::post('/store', [SchoolController::class, 'store'])->middleware('permission:create_school')->name('store-school');
+            Route::post('/store-detail/{id}', [SchoolDetailController::class, 'store'])->middleware('permission:read_school')->name('store-detail');
+            Route::get('/edit/{id}', [SchoolController::class, 'edit'])->middleware('permission:update_school')->name('edit-school');
+            Route::get('/edit-detail/{id}', [SchoolDetailController::class, 'edit'])->middleware('permission:update_school')->name('edit-detail');
+            Route::post('/update/{id}', [SchoolController::class, 'update'])->middleware('permission:update_school')->name('update-school');
+            Route::post('/update-detail/{id}', [SchoolDetailController::class, 'update'])->middleware('permission:update_school')->name('update-detail');
+            Route::delete('/destroy/{id}', [SchoolController::class, 'destroy'])->middleware('permission:delete_school')->name('destroy-school');
         });
 
     Route::prefix('calculate')->group(function (){
         Route::get('/', [CalculateController::class, 'index'])->name('calculate');
         Route::get('/weighting/{id}', [AhpController::class, 'index'])->name('weighting');
-        Route::post('/process', [AhpController::class, 'calculate'])->name('process');
+        Route::post('/process/{id}', [AhpController::class, 'calculate'])->name('process');
         Route::get('/alternate/{id}', [ArasController::class, 'index'])->name('alternate');
         Route::post('/store/{id}', [ArasController::class, 'store'])->name('store-aras');
         Route::get('/direction/{id}', [ArasController::class, 'direction'])->name('direction');
